Convert image storage to plain filesystem

+++ server/models/image.js
@@ -1,38 +1,76 @@
+const fs = require('fs');
+const async = require('async');
+const multer = require('multer');
 const DBTools = require('../middleware/db');
 
-const db = new DBTools('fs.files');
+const db = new DBTools('images');
+
+const IMAGE_TYPES = [
+  'image/png',
+  'image/jpeg',
+  'image/gif',
+  'image/bmp',
+  'image/gif',
+  'image/tiff'
+];
+
+const storage = multer.diskStorage({
+  destination: (req, file, cb) => {
+    // Multer is lazy...
+    if (!fs.existsSync(req.session.uid)) {
+      fs.mkdirSync(req.session.uid);
+    }
+    cb(null, req.session.uid);
+  }
+});
+
+function fileFilter(req, file, cb) {
+  return cb(null, IMAGE_TYPES.some(type => file.mimetype === type));
+}
 
 module.exports = {
-  add(uid, imageId, next) {
-    // The file should already exist thanks to multer
-    db.updateOne({ _id: db.getId(imageId) }, { uid, shared: false }, (success) => {
-      if (!success) {
-        next('Failure adding user id to image');
-      } else {
-        next(null);
-      }
-    });
+  uploadMiddleware: multer({ storage, fileFilter }).array('images'),
+
+  addMany(images, cb) {
+    db.insertMany(images, cb);
   },
 
   get(uid, id, next) {
     db.findOne({ _id: db.getId(id) }, (doc) => {
       if (!doc) {
-        next('cannot find image', null);
+        next(404, null);
       } else if (doc.uid !== uid && !doc.shared) {
-        next(401, null); // using a number here prevents buggy string comp
+        next(401, null);
       } else {
-        next(null, {
-          file: db.readFile(id),
-          contentType: doc.contentType
-        });
+        next(null, doc);
       }
     });
   },
 
   remove(uid, id, next) {
-    db.removeOne({ _id: db.getId(id), uid }, (removed) => {
-      if (removed) return next();
-      return next('cannot find image, or invalid permissions');
+    module.exports.get(uid, id, (err, file) => {
+      if (err) {
+        return next('cannot find image, or invalid permissions');
+      }
+      if (file.refs - 1 === 0) {
+        return db.removeOne({ _id: db.getId(id), uid }, (removed) => {
+          if (!removed) {
+            return next('failed to remove image');
+          }
+          fs.unlinkSync(file.path);
+          return next();
+        });
+      }
+      return db.updateRaw(
+        { _id: db.getId(id), uid },
+        { $inc: { refs: -1 } },
+        (updated) => {
+          if (!updated) {
+            return next('failed to remove image');
+          }
+          return next();
+        }
+      );
     });
   },
 
@@ -60,6 +98,16 @@ module.exports = {
   },
 
   purgeUserImages: (uid, cb) => {
-    db.removeMany({ uid }, () => cb());
+    db.findMany({ uid }, images =>
+      async.each(
+        images,
+        (image, next) =>
+          fs.unlink(image.path, next),
+        (err) => {
+          if (err) cb('failed to purge user');
+          else db.removeMany({ uid }, () => cb());
+        }
+      )
+    );
   }
 };
